aula 12/11

+++ em aula/expressoes/Financiamento_1.java
@@ -2,7 +2,7 @@ import java.lang.Math;
 // <<class definition>>
 class Financiamento_1 {
     // <<function-method definition>>
-    int pm_1(int v, int i, int n) {
+    double pm_1(double v, double i, int n) {
         double k = Math.pow(1 + i, n);
         double cf = (k * i) / (k - 1);
         return v * cf;
